Moved files service functions to separate files
