+++ src/services/files/mod.rs
@@ -1,0 +1,3 @@
+pub mod delete;
+pub mod get;
+pub mod insert;
